Release workflow (#544)
This is a modified version of a local script i've been using for a few
releases. Hasn't been tested as an action yet so we'll need a new
release to validate.

### New workflows for release management:

#### Prepare Release Workflow:
* Added `.github/workflows/prepare-release.yml` to automate the creation
of draft releases, release branches, and pull requests for new versions.
Includes support for potential release/hotfix base branches in the case
of a backport fix. To create a backport, we'll just need to create a
branch for the target base version (release/v1.x) and set the base
branch for the action.

#### Publish Release Workflow:
* Added `.github/workflows/publish-release.yml` to handle the
publication of merged release pull requests, including publishing to
npm, updating GitHub release notes

#### Update Fixed Issues Workflow:
* Added `.github/workflows/updated-fixed-issues.yml` to comment on fixed
issues.

---------

Signed-off-by: Paul Sachs <psachs@buf.build>